app/requirements.txt
Flask
flask-mongoengine
requests
flask-swagger
flask-cors